Pull out string based utilities to dedicated file
