+++ src/Utilities/Utilities.h
@@ -43,6 +43,8 @@ namespace Utilities
 
     bool isSubtractionSafe(int64_t currentValue, uint64_t transferAmount);
 
+    bool parseDaemonAddressFromString(std::string &host, uint16_t &port, std::string address);
+
     /* Verify that a + b will not overflow when added. */
     /* 2 positive numbers - should always get greater (or equal) when summed. */
     /* Any negative numbers - should always get smaller (or equal) when summed. */
