+++ src/zedwallet++/GetInput.cpp
@@ -12,9 +12,11 @@
 
 #include "linenoise.hpp"
 
-#include <Utilities/FormatTools.h>
-
 #include <Utilities/ColouredMsg.h>
+#include <Utilities/FormatTools.h>
+#include <Utilities/String.h>
+#include <Utilities/Utilities.h>
+
 #include <zedwallet++/Commands.h>
 #include <zedwallet++/Utilities.h>
 
@@ -89,7 +91,7 @@ std::string getInput(
     }
 	
     /* Remove any whitespace */
-    Common::trim(command);
+    Utilities::trim(command);
 
     if (command != "")
     {
@@ -116,7 +118,7 @@ std::string getAddress(
             return "cancel";
         }
 
-        Common::trim(address);
+        Utilities::trim(address);
 
         /* \n == no-op */
         if (address == "")
@@ -160,7 +162,7 @@ std::string getPaymentID(
             return "cancel";
         }
 
-        Common::trim(paymentID);
+        Utilities::trim(paymentID);
 
         if (paymentID == "cancel" && cancelAllowed)
         {
@@ -201,7 +203,7 @@ std::string getHash(
             return "cancel";
         }
 
-        Common::trim(hash);
+        Utilities::trim(hash);
 
         if (hash == "cancel" && cancelAllowed)
         {
@@ -244,10 +246,10 @@ std::tuple<bool, uint64_t> getAmountToAtomic(
             continue;
         }
 
-        Common::trim(amountString);
+        Utilities::trim(amountString);
 
         /* If the user entered thousand separators, remove them */
-        ZedUtilities::removeCharFromString(amountString, ',');
+        Utilities::removeCharFromString(amountString, ',');
 
         if (amountString == "cancel" && cancelAllowed)
         {
@@ -276,7 +278,7 @@ std::tuple<bool, uint64_t> getAmountToAtomic(
         }
 
         /* Remove the decimal place, so we can parse it as an atomic amount */
-        ZedUtilities::removeCharFromString(amountString, '.');
+        Utilities::removeCharFromString(amountString, '.');
 
         /* Pad the string with 0's at the end, so 123 becomes 12300, so we 
            can parse it as an atomic amount. 123.45 parses as 12345. */
@@ -330,9 +332,9 @@ std::tuple<std::string, uint16_t> getDaemonAddress()
             return {host, port};
         }
 
-        Common::trim(address);
+        Utilities::trim(address);
 
-        if (!ZedUtilities::parseDaemonAddressFromString(host, port, address))
+        if (!Utilities::parseDaemonAddressFromString(host, port, address))
         {
             std::cout << WarningMsg("\nInvalid daemon address! Try again.\n");
             continue;
