+++ src/zedwallet++/Utilities.cpp
@@ -15,6 +15,8 @@
 #include <iostream>
 
 #include <Utilities/ColouredMsg.h>
+#include <Utilities/String.h>
+
 #include <zedwallet++/PasswordContainer.h>
 
 namespace ZedUtilities
@@ -108,7 +110,7 @@ uint64_t getScanHeight()
         std::getline(std::cin, stringHeight);
 
         /* Remove commas so user can enter height as e.g. 200,000 */
-        removeCharFromString(stringHeight, ',');
+        Utilities::removeCharFromString(stringHeight, ',');
 
         if (stringHeight == "")
         {
@@ -129,102 +131,6 @@ uint64_t getScanHeight()
                       << WarningMsg("a number!") << std::endl << std::endl;
         }
     }
-}
-
-/* Erases all instances of c from the string. E.g. 2,000,000 becomes 2000000 */
-void removeCharFromString(std::string &str, const char c)
-{
-    str.erase(std::remove(str.begin(), str.end(), c), str.end());
-}
-
-/* Trims any whitespace from left and right */
-void trim(std::string &str)
-{
-    rightTrim(str);
-    leftTrim(str);
-}
-
-void leftTrim(std::string &str)
-{
-    std::string whitespace = " \t\n\r\f\v";
-
-    str.erase(0, str.find_first_not_of(whitespace));
-}
-
-void rightTrim(std::string &str)
-{
-    std::string whitespace = " \t\n\r\f\v";
-
-    str.erase(str.find_last_not_of(whitespace) + 1);
-}
-
-/* Checks if str begins with substring */
-bool startsWith(const std::string &str, const std::string &substring)
-{
-    return str.rfind(substring, 0) == 0;
-}
-
-std::vector<std::string> split(const std::string& str, char delim = ' ')
-{
-    std::vector<std::string> cont;
-    std::stringstream ss(str);
-    std::string token;
-
-    while (std::getline(ss, token, delim))
-    {
-        cont.push_back(token);
-    }
-
-    return cont;
-}
-
-std::string removePrefix(const std::string &str, const std::string &prefix)
-{
-    const size_t removePos = str.rfind(prefix, 0);
-
-    if (removePos == 0)
-    {
-        /* erase is in place */
-        std::string copy = str;
-
-        copy.erase(0, prefix.length());
-
-        return copy;
-    }
-
-    return str;
-}
-
-bool parseDaemonAddressFromString(std::string &host, uint16_t &port, std::string address)
-{
-    /* Lets users enter url's instead of host:port */
-    address = removePrefix(address, "https://");
-    address = removePrefix(address, "http://");
-
-    std::vector<std::string> parts = split(address, ':');
-
-    if (parts.empty())
-    {
-        return false;
-    }
-    else if (parts.size() >= 2)
-    {
-        try
-        {
-            host = parts.at(0);
-            port = std::stoi(parts.at(1));
-            return true;
-        }
-        catch (const std::invalid_argument &)
-        {
-            return false;
-        }
-    }
-
-    host = parts.at(0);
-    port = CryptoNote::RPC_DEFAULT_PORT;
-
-    return true;
 }
 
 } // namespace
